wip: split dev/full image for size reasons, start publishing them

+++ docker/benchbase/devcontainer/build-in-container.sh
@@ -5,10 +5,12 @@
 set -eu -o pipefail
 
 BENCHBASE_PROFILES=${BENCHBASE_PROFILES:-cockroachdb mariadb mysql postgres spanner phoenix sqlserver sqlite}
+CLEAN_BUILD=${CLEAN_BUILD:-true}    # true, false, pre, post
+TEST_TARGET=${TEST_TARGET:-test}    # or empty
 
 cd /benchbase
 
-function build_profile {
+function build_profile() {
     local profile="$1"
     # Build in a separate directory so we can do it in parallel.
     rm -rf profiles/$profile/
@@ -21,7 +23,7 @@ function build_profile {
     rm -rf profiles/$profile/data/ && ln -s ../../data profiles/$profile/data
 }
 
-function test_profile_build {
+function test_profile_build() {
     local profile="$1"
     if [ "$(readlink -f profiles/$profile/data)" != "/benchbase/data" ]; then
         echo "ERROR: profiles/$profile/data is not /benchbase/data." >&2
@@ -29,11 +31,23 @@ function test_profile_build {
     fi
 }
 
-function clean_profile_build {
+function clean_profile_build() {
     local profile="$1"
-    mvn -B --file pom.xml -D buildDirectory=target/$profile clean
-    rm -rf target/$profile
+    if [ -d target/$profile ]; then
+        mvn -B --file pom.xml -D buildDirectory=target/$profile clean
+        rm -rf target/$profile
+    fi
 }
+
+if [ "$CLEAN_BUILD" == 'true' ] || [ "$CLEAN_BUILD" == 'pre' ]; then
+    for profile in ${BENCHBASE_PROFILES}; do
+        clean_profile_build "$profile" &
+    done
+fi
+wait
+
+# Make sure that we've build the base stuff first before we start
+mvn -T 2C -B --file pom.xml process-resources compile $TEST_TARGET
 
 for profile in ${BENCHBASE_PROFILES}; do
     build_profile "$profile" &
@@ -45,7 +59,9 @@ for profile in ${BENCHBASE_PROFILES}; do
     test_profile_build "$profile" || exit 1
 done
 
-for profile in ${BENCHBASE_PROFILES}; do
-    clean_profile_build "$profile" &
-done
-wait
+if [ "$CLEAN_BUILD" == 'true' ] || [ "$CLEAN_BUILD" == 'post' ]; then
+    for profile in ${BENCHBASE_PROFILES}; do
+        clean_profile_build "$profile" &
+    done
+    wait
+fi
